requirements.txt
fire==0.5.0
requests==2.31.0
pillow==10.1.0
psycopg2-binary==2.9.9